requirements.txt
flask
flask-cors
pyautogui
opencv-python
fpdf2
psutil
numpy